+++ src/pages/Domain.tsx
@@ -1,4 +1,4 @@
-import Layout from "../Layout";
+import Layout from "../components/Layout";
 
 const Domain = ({ domain }: { domain: string }) => (
   <Layout pageTitleSuffix={`Är domänen ${domain} ledig?`}>
